Squashed commit of the following:

commit 4ab061afb04cc904a71faf19eadd5885acf8e58b
Merge: 5f95dd6 adb556e
Author: ringe <ringe@mecom.de>
Date:   Thu Aug 24 08:20:40 2017 +0200

    Merge branch 'master' of github.com:faburem/dokuwiki-docker into faburem-master

commit adb556ec60eb992837e21d4d0fbea08859e4346b
Author: Fabian Kromer <fabian.kromer@gmail.com>
Date:   Fri Aug 11 16:41:11 2017 +0200

    exporting the plugins folder as separate volume and make sure it works with persistent volumes

commit 278a9b73f321c058d72b9ed854f433caa7277290
Author: Fabian Kromer <fabian.kromer@gmail.com>
Date:   Fri Aug 11 15:45:13 2017 +0200

    making sure to only overwrite configs the first time the container is run to enable better support for persistent volumes

+++ docker-entrypoint.sh
@@ -18,14 +18,23 @@ set +e
 
 # move directories out of web root
 mv $BASE_DIR/data/* /dokuwiki-data/
-mv $BASE_DIR/conf/* /dokuwiki-conf/
+mv $BASE_DIR/lib/plugins/* /dokuwiki-plugins/
+rm -rf $BASE_DIR/lib/plugins/
+ln -s /dokuwiki-plugins $BASE_DIR/lib/plugins
 
+#do not replace the configs if we find existing ones
+if [ ! -f "/dokuwiki-conf/local.php" ]; then
+ echo "creating initial config ...";
+ mv $BASE_DIR/conf/* /dokuwiki-conf/
+fi
 # create preload.php
 cat << EOF > $BASE_DIR/inc/preload.php
 <?php
 define('DOKU_CONF','/dokuwiki-conf/');
 EOF
 
+#only change the config if we don't find existing ones
+if [ ! -f "/dokuwiki-conf/local.php" ]; then
 # add savedir option to local configuration
 cat << EOF > /dokuwiki-conf/local.php
 <?php
@@ -44,7 +53,7 @@ EOF
 cat << EOF > /dokuwiki-conf/users.auth.php
 admin:21232f297a57a5a743894a0e4a801fc3:admin:admin@localhost:admin,users,devel,support
 EOF
-
+fi
 # run dokuwiki
 php -S 0.0.0.0:80 -t $BASE_DIR/
